+++ client/components/edit-page/results.js
@@ -1,33 +1,34 @@
 import React from 'react'
 
 const Results = ({testCases, outputs}) => {
-  testCases = testCases.trim().split('\n')
+  testCases = testCases ? testCases.trim().split('\n') : ['']
   if (typeof outputs === 'string') {
     return outputs
   } else if (outputs.length === 0) {
     return 'No Output'
   }
-  const type = elem =>
+
+  const type = text =>
     Object.prototype.toString
-      .call(elem)
+      .call(text)
       .slice(8, -1)
       .toLowerCase()
-  const getOutput = output => {
-    const outputType = type(output)
-    if (outputType === 'null' || outputType === 'undefined') {
-      return outputType
-    } else if (outputType === 'boolean') {
-      return output ? 'true' : 'false'
+  const renderText = text => {
+    if (['null', 'boolean'].includes(type(text))) {
+      return JSON.stringify(text)
+    } else if (type(text) === 'undefined') {
+      return 'undefined'
     } else {
-      return output
+      return text
     }
   }
+
   return (
     <ul>
       {testCases.map((input, i) => (
         <li key={i}>
-          <div>Input: {input}</div>
-          <div>Your Output: {getOutput(outputs[i])}</div>
+          <div>Input: {renderText(input)}</div>
+          <div>Your Output: {renderText(outputs[i])}</div>
         </li>
       ))}
     </ul>
