+++ client/components/edit-page/test-cases.js
@@ -1,12 +1,40 @@
 import React from 'react'
+import {withStyles} from '@material-ui/core/styles'
+import TextField from '@material-ui/core/TextField'
 
-const TestCases = ({onChange, testCases}) => {
+const styles = theme => ({
+  container: {
+    display: 'flex',
+    flexWrap: 'wrap'
+  },
+  textField: {
+    marginLeft: theme.spacing.unit,
+    marginRight: theme.spacing.unit
+  },
+  dense: {
+    marginTop: 16
+  },
+  menu: {
+    width: '80%',
+    height: '80%'
+  }
+})
+
+const TestCases = ({onChange, testCases, classes}) => {
+  console.log('new test case')
   return (
-    <textarea
-      rows="20"
-      cols="50"
+    <TextField
+      id="outlined-textarea"
+      label="Input"
+      placeholder="Make sure you write each test case on a separate line."
+      multiline
+      className={classes.textField}
+      margin="normal"
+      variant="outlined"
+      fullWidth={true}
       name="testCases"
       onChange={event => {
+        console.log(event)
         const {target} = event
         onChange(target.name, target.value)
       }}
@@ -15,4 +43,4 @@ const TestCases = ({onChange, testCases}) => {
   )
 }
 
-export default TestCases
+export default withStyles(styles)(TestCases)
