+++ client/components/navbar.js
@@ -1,75 +1,182 @@
 import React from 'react'
+import AppBar from '@material-ui/core/AppBar'
+import Toolbar from '@material-ui/core/Toolbar'
+import Typography from '@material-ui/core/Typography'
+import {withStyles} from '@material-ui/core/styles'
+import Button from '@material-ui/core/Button'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
+import List from '@material-ui/core/List'
+import ListItem from '@material-ui/core/ListItem'
+import ListItemText from '@material-ui/core/ListItemText'
+import SwipeableDrawer from '@material-ui/core/SwipeableDrawer'
+import BreakpointMedia from 'react-media-material-ui/BreakpointMedia'
 import {Link} from 'react-router-dom'
-import {logout} from '../store'
-import HamburgerMenu from 'react-hamburger-menu'
+import MenuIcon from '@material-ui/icons/Menu'
 
-const Navbar = ({isLoggedIn, logout, menuOpen, toggleMenu}) => {
-  return (
-    <div>
-      <nav id="nav-bar">
-        <nav role="navigation">
-          <div id="menuToggle">
-            <input type="checkbox" />
-
-            <span />
-            <span />
-            <span />
-
-            <ul id="menu">
-              {isLoggedIn ? (
-                <div id="menu-link-list">
-                  {/* The navbar will show these links after you log in */}
-                  <button onClick={logout}>Logout</button>
-                </div>
-              ) : (
-                <div id="menu-link-list">
-                  {/* The navbar will show these links before you log in */}
-                  <Link to="/login">Login</Link>
-                  <Link to="/signup">Sign Up</Link>
-                </div>
-              )}
-            </ul>
-          </div>
-        </nav>
-        <div id="title-container">
-          <Link to="/">
-            <h1 id="title">Kiwi</h1>
-          </Link>
-        </div>
-      </nav>
-      <hr />
-    </div>
-  )
-}
-
-/**
- * CONTAINER
- */
-const mapState = state => {
-  return {
-    isLoggedIn: !!state.user.id
+const style = theme => ({
+  appBar: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    zIndex: theme.zIndex.drawer + 1,
+    transition: theme.transitions.create(['width', 'margin'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    })
+  },
+  appBarSpacer: {
+    minHeight: '30px'
+  },
+  content: {
+    flexGrow: 1,
+    padding: theme.spacing.unit * 2,
+    height: '100vh',
+    overflow: 'auto'
+  },
+  toolbar: {
+    display: 'flex',
+    flexDirection: 'row',
+    alignItems: 'flexEnd',
+    justifyContent: 'space-around'
+  },
+  button: {
+    margin: theme.spacing.unit
+  },
+  input: {
+    display: 'none'
+  },
+  dropdown: {
+    width: 'auto'
+  },
+  titleLink: {
+    paddingLeft: '20px',
+    flexGrow: 4
   }
-}
+})
 
-const mapDispatch = dispatch => {
-  return {
-    logout() {
-      dispatch(logout())
+class Navbar extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      open: false
     }
   }
+  toggleDrawer = () => () => {
+    const status = this.state.open
+    this.setState({
+      open: !status
+    })
+  }
+
+  render() {
+    const {classes, isLoggedIn} = this.props
+    const loggedIn = [
+      {
+        text: 'Home',
+        link: '/home'
+      },
+      {
+        text: 'Evaluate',
+        link: '/evaluate'
+      },
+      {
+        text: 'FAQs',
+        link: '/faq'
+      },
+      {
+        text: 'Logout',
+        link: '/logout'
+      }
+    ]
+    const loggedOut = [
+      {
+        text: 'Evaluate',
+        link: '/evaluate'
+      },
+      {
+        text: 'FAQs',
+        link: '/faq'
+      },
+      {
+        text: 'Login',
+        link: '/login'
+      },
+      {
+        text: 'Sign Up',
+        link: '/signup'
+      }
+    ]
+
+    const currentList = isLoggedIn ? loggedIn : loggedOut
+    return (
+      <div>
+        <AppBar position="static" className={classes.appBar}>
+          <div className={classes.titleLink}>
+            <Link to="/" className={classes.titleLink}>
+              <img src="/logo.png" alt="Kiwi" />
+            </Link>
+          </div>
+          <div className={classes.navlinks}>
+            <Toolbar className={classes.toolbar}>
+              <BreakpointMedia max="xs">
+                <div>
+                  <SwipeableDrawer
+                    anchor="top"
+                    open={this.state.open}
+                    onClose={this.toggleDrawer()}
+                    onOpen={this.toggleDrawer()}
+                  >
+                    <div
+                      tabIndex={0}
+                      role="button"
+                      onClick={this.toggleDrawer()}
+                      onKeyDown={this.toggleDrawer()}
+                    >
+                      <div className={classes.fullList}>
+                        <List>
+                          {currentList.map(item => (
+                            <Link to={item.link} key={item.text}>
+                              <ListItem button>
+                                <ListItemText primary={item.text} />
+                              </ListItem>
+                            </Link>
+                          ))}
+                        </List>
+                      </div>
+                    </div>
+                  </SwipeableDrawer>
+
+                  <Button onClick={this.toggleDrawer()}>
+                    <MenuIcon />
+                  </Button>
+                </div>
+              </BreakpointMedia>
+              <BreakpointMedia min="sm">
+                <div>
+                  {currentList.map(item => (
+                    <Link to={item.link} key={item.text}>
+                      <Button>{item.text}</Button>
+                    </Link>
+                  ))}
+                </div>
+              </BreakpointMedia>
+            </Toolbar>
+          </div>
+        </AppBar>
+        <div className={classes.appBarSpacer} />
+      </div>
+    )
+  }
 }
 
-export default connect(
-  mapState,
-  mapDispatch
-)(Navbar)
-
-/**
- * PROP TYPES
- */
 Navbar.propTypes = {
-  toggleMenu: PropTypes.func.isRequired,
-  isLoggedIn: PropTypes.bool.isRequired
+  classes: PropTypes.object.isRequired
 }
+
+const mapState = state => ({
+  isLoggedIn: !!state.user.id
+})
+
+export default connect(mapState)(withStyles(style)(Navbar))
